separate Course into own module
