+++ part2/courseinfo/src/App.jsx
@@ -1,72 +1,60 @@
-const Header = ({ course }) => {
-  return (
-    <h1>{course}</h1>
-  );
-};
-
-const Part = ({ name, number }) => {
-  return (
-    <p>
-      {name} {number}
-    </p>
-  );
-};
-
-const Content = ({ parts }) => {
-  return (
-    <div>
-      {parts.map((part, i) => (
-        <Part name={part.name} number={part.exercises} key={part.id}/>
-      ))}
-    </div>
-  );
-};
-
-const Total = ({ parts }) => {
-  const total = parts.reduce((total, part) => {
-    return total + part.exercises;
-  }, 0)
-
-  return (
-    <b>total of {total} exercises</b>
-  );
-};
-
-const Course = ({ course }) => { 
-  return (
-    <div>
-      <Header course={course.name} />
-      <Content parts={course.parts}/>
-      <Total parts={course.parts} />
-    </div>
-  );
-
- };
+import Course from './Course.jsx';
 
 const App = () => {
-  const course = {
-    id: 1,
-    name: 'Half Stack application development',
-    parts: [
-      {
-        name: 'Fundamentals of React',
-        exercises: 10,
-        id: 1
-      },
-      {
-        name: 'Using props to pass data',
-        exercises: 7,
-        id: 2
-      },
-      {
-        name: 'State of a component',
-        exercises: 14,
-        id: 3
-      }
-    ]
-  };
+  const courses = [
+    {
+      name: 'Half Stack application development',
+      id: 1,
+      parts: [
+        {
+          name: 'Fundamentals of React',
+          exercises: 10,
+          id: 1
+        },
+        {
+          name: 'Using props to pass data',
+          exercises: 7,
+          id: 2
+        },
+        {
+          name: 'State of a component',
+          exercises: 14,
+          id: 3
+        },
+        {
+          name: 'Redux',
+          exercises: 11,
+          id: 4
+        }
+      ]
+    }, 
+    {
+      name: 'Node.js',
+      id: 2,
+      parts: [
+        {
+          name: 'Routing',
+          exercises: 3,
+          id: 1
+        },
+        {
+          name: 'Middlewares',
+          exercises: 7,
+          id: 2
+        }
+      ]
+    }
+  ]
 
-  return <Course course={course} />;
+  return (
+    <div>
+      <h1>Web Development Curriculum</h1>
+      { courses.map(course => (
+        <Course course={course} key={course.id}/>
+      ))
+      }
+    </div>
+  );
 };
 
 export default App;
